+++ MainWindow.xaml.cs
@@ -10,6 +10,7 @@ namespace BauphysikToolWPF
     {
         public static ListBox navigationMenuListBox;
 
+        // Saves MainWindow Instance here
         public static Window Main;
 
         public MainWindow()
@@ -18,8 +19,20 @@ namespace BauphysikToolWPF
             navigationMenuListBox = this.NavigationMenuListBox;
             Main = this;
         }
+        
+        // The Pages a user can inteact with
+        public enum NavigationContent
+        {
+            LandingPage,
+            SetupLayer,
+            SetupEnv,
+            TemperatureCurve,
+            GlaserCurve,
+            Default = LandingPage
+        }
         public static void SetPage(string page)
         {
+            
             navigationMenuListBox.SelectedItem = page; // MainWindow.xaml Binding changes the ContentPage based on the SelectedItem string
         }
     }
