+++ UI/FO1_SetupLayer.xaml.cs
@@ -10,9 +10,9 @@ using System.Windows.Input;
 namespace BauphysikToolWPF.UI
 {
     /// <summary>
-    /// Interaktionslogik für FO1_Setup.xaml
+    /// Interaktionslogik für FO1_SetupLayer.xaml
     /// </summary>
-    public partial class FO1_Setup : UserControl
+    public partial class FO1_SetupLayer : UserControl
     {
         // Class Variables - Belongs to the Class-Type itself and stay the same
 
@@ -26,7 +26,7 @@ namespace BauphysikToolWPF.UI
         private int currentElementId { get; } = FO0_LandingPage.SelectedElementId;
 
         // (Instance-) Contructor - when 'new' Keyword is used to create class (e.g. when toggling pages via menu navigation)
-        public FO1_Setup()
+        public FO1_SetupLayer()
         {   
             // UI Elements in backend only accessible AFTER InitializeComponent() was executed
             InitializeComponent(); // Initializes xaml objects -> Calls constructors for all referenced Class Bindings in the xaml (from DataContext, ItemsSource etc.)                                                    
@@ -102,51 +102,6 @@ namespace BauphysikToolWPF.UI
 
         // UI Methods
 
-        // When User sets Custom Value for EnvVars
-        private void numericData_PreviewTextInput(object sender, TextCompositionEventArgs e)
-        {
-            //Handle the input
-            string userInput = e.Text;
-            Regex regex = new Regex("[^0-9,-]+"); //regex that matches disallowed text
-            e.Handled = regex.IsMatch(e.Text);
-
-            // only allow one decimal point
-            if (userInput == "." && (sender as TextBox).Text.IndexOf('.') > -1)
-            {
-                e.Handled = true;
-            }
-
-            //set new value as UserSaved Data
-            /*switch ((sender as TextBox).Name)
-            {
-                case "Ti_Input":
-                    UserSaved.Ti = Convert.ToDouble(Ti_Input.Text + userInput);
-                    Ti_ComboBox.SelectedIndex = -1; // set empty selection
-                    return;
-                case "Te_Input":
-                    UserSaved.Te = Convert.ToDouble(Te_Input.Text + userInput);
-                    Te_ComboBox.SelectedIndex = -1; // set empty selection
-                    return;
-                case "Rsi_Input":
-                    UserSaved.Rsi = Convert.ToDouble(Rsi_Input.Text + userInput);
-                    Rsi_ComboBox.SelectedIndex = -1; // set empty selection
-                    return;
-                case "Rse_Input":
-                    UserSaved.Rse = Convert.ToDouble(Rse_Input.Text + userInput);
-                    Rse_ComboBox.SelectedIndex = -1; // set empty selection
-                    return;
-                case "Rel_Fi_Input":
-                    UserSaved.Rel_Fi = Convert.ToDouble(Rel_Fi_Input.Text + userInput);
-                    Rel_Fi_ComboBox.SelectedIndex = -1; // set empty selection
-                    return;
-                case "Rel_Fe_Input":
-                    UserSaved.Rel_Fe = Convert.ToDouble(Rel_Fe_Input.Text + userInput);
-                    Rel_Fe_ComboBox.SelectedIndex = -1; // set empty selection
-                    return;
-                default: throw new ArgumentException("Could not assign value");
-            }*/
-        }
-
         // Redraw on every Click on Canvas to show current selected Layer with blue border
         private void layers_ListView_SelectionChanged(object sender, SelectionChangedEventArgs e)
         {
@@ -169,7 +124,7 @@ namespace BauphysikToolWPF.UI
             new DrawLayerCanvas(layers_Canvas, layers_ListView.ItemsSource as List<Layer>);
         }
 
-        // Save current canvas as image, just before closing FO1_Setup Page
+        // Save current canvas as image, just before closing FO1_SetupLayer Page
         // 'Unloaded' event was called after FO0 Initialize();
         private void UserControl_IsVisibleChanged(object sender, DependencyPropertyChangedEventArgs e)
         {
