+++ UI/FO3_Moisture.xaml.cs
@@ -13,12 +13,12 @@ namespace BauphysikToolWPF.UI
         {
             // Save computation time and only recalculate if needed
             // Only if Element, Layers or EnvVars are not set or have changed: update class variables.
-            if (FO1_Setup.RecalculateGlaser == true)
+            if (FO1_SetupLayer.RecalculateGlaser == true)
             {
                 GlaserCalculation = new GlaserCalc(); // for FO3_ViewModel
 
                 // Reset Recalculate Flag
-                FO1_Setup.RecalculateGlaser = false;
+                FO1_SetupLayer.RecalculateGlaser = false;
             }
             InitializeComponent();
         }
